- Finally finished live-mode system monitoring branch. It uses psutil to receive CPU and RAM related data. - Added ram use percentage row to packet and DB. - Changed reading value column from Int to Float. - Moved SQLite3 database file to respective directory. It contains about 400 rows of device status logs (emulation) and some basic records for all objects - just for demonstration purposes.

+++ server/database/models/device/device_monitor_readings.py
@@ -1,7 +1,7 @@
 """
 DeviceSystemMonitorReadings table model and migration/creation events container
 """
-from sqlalchemy import Column, Integer, DateTime, ForeignKey
+from sqlalchemy import Column, Integer, DateTime, ForeignKey, Float
 from sqlalchemy.orm import relationship
 
 from server.web import db as database
@@ -16,5 +16,5 @@ class DeviceSystemMonitorReadings(database.Model):
     device_id = Column(Integer())
     reading_id = Column(Integer(), ForeignKey("device_system_monitor_reading_types.id"))
     reading_type = relationship("DeviceSystemMonitorReadingTypes")
-    value = Column(Integer())
+    value = Column(Float())
     reported_at = Column(DateTime())
